+++ pytsite/google/res/assets/js/maps.js
@@ -18,7 +18,7 @@ pytsite.google.maps.Map = function (mapNode, options) {
 
     // Options
     self.options = {
-        // Read https://developers.google.com/maps/documentation/javascript/controls
+        // https://developers.google.com/maps/documentation/javascript/controls
         center: new google.maps.LatLng(50.45, 30.52),
         zoom: 18,
 
@@ -41,7 +41,7 @@ pytsite.google.maps.Map = function (mapNode, options) {
     };
     $.extend(true, self.options, options);
 
-    self.map = new google.maps.Map(mapNode, self.options);
+    self.mapObject = new google.maps.Map(mapNode, self.options);
     self.mapNode = mapNode;
     self.currentPosition = self.options.center;
     self.markers = [];
@@ -53,7 +53,7 @@ pytsite.google.maps.Map = function (mapNode, options) {
         // Create and add button
         var btn = $('<a class="pytsite-google-map-control center-map hidden" href="#">' +
             '<i class="fa fa-fw fa-3x fa-crosshairs"></i></a>');
-        self.map.controls[google.maps.ControlPosition.RIGHT_TOP].push(btn[0]);
+        self.mapObject.controls[google.maps.ControlPosition.RIGHT_TOP].push(btn[0]);
 
         // Center the map and hide the button after click on it
         btn.click(function (e) {
@@ -66,7 +66,7 @@ pytsite.google.maps.Map = function (mapNode, options) {
         self.mapCenterControl = btn;
 
         // Show the button on map dragging
-        self.map.addListener('drag', function () {
+        self.mapObject.addListener('drag', function () {
             if (self.mapCenterControl) {
                 self.mapCenterControl.removeClass('hidden');
             }
@@ -77,7 +77,7 @@ pytsite.google.maps.Map = function (mapNode, options) {
     if (self.options.currentPositionMarker) {
         $.extend(self.options.currentPositionMarkerOptions, {
             position: self.currentPosition,
-            map: self.map
+            map: self.mapObject
         });
         self.currentPositionMarker = new google.maps.Marker(self.options.currentPositionMarkerOptions);
     }
@@ -125,16 +125,20 @@ pytsite.google.maps.Map = function (mapNode, options) {
         );
     }
 
-    // === Center map to the current position ===
+    /**
+     * Center map to its current position.
+     */
     self.setCenterToCurrentPosition = function () {
-        self.map.setCenter(self.currentPosition);
+        self.mapObject.setCenter(self.currentPosition);
 
         if (self.options.mapCenterControl) {
             self.mapCenterControl.addClass('hidden');
         }
     };
 
-    // === Adjusts map viewport to make all markers visible ===
+    /**
+     * Adjusts map viewport to make all markers visible.
+     */
     self.fitBounds = function () {
         var markerBounds = new google.maps.LatLngBounds();
         markerBounds.extend(self.currentPositionMarker.getPosition());
@@ -143,12 +147,14 @@ pytsite.google.maps.Map = function (mapNode, options) {
             markerBounds.extend(self.markers[i].getPosition());
         }
 
-        self.map.fitBounds(markerBounds);
-        if (self.map.getZoom() > self.options.zoom)
-            self.map.setZoom(self.options.zoom);
+        self.mapObject.fitBounds(markerBounds);
+        if (self.mapObject.getZoom() > self.options.zoom)
+            self.mapObject.setZoom(self.options.zoom);
     };
 
-    // === Get current location ===
+    /**
+     * Update current coordinates.
+     */
     self.updateCurrentPosition = function () {
         var defer = $.Deferred();
 
@@ -170,14 +176,18 @@ pytsite.google.maps.Map = function (mapNode, options) {
         return defer;
     };
 
-    // === Clear drawn routes ===
+    /**
+     * Clear drawn routes.
+     */
     self.clearRoutes = function () {
         if (self.options.directionsService) {
             self.directionsRenderer.setMap(null);
         }
     };
 
-    // === Draw route ===
+    /**
+     * Draw a route.
+     */
     self.drawRoute = function (origin, destination, travelMode) {
         var request = {
             origin: origin,
@@ -188,7 +198,7 @@ pytsite.google.maps.Map = function (mapNode, options) {
         // Making request to Google Directions Service
         self.directionsService.route(request, function (result, status) {
             if (status == google.maps.DirectionsStatus.OK) {
-                self.directionsRenderer.setMap(self.map);
+                self.directionsRenderer.setMap(self.mapObject);
                 self.directionsRenderer.setDirections(result);
             }
             else {
@@ -197,11 +207,34 @@ pytsite.google.maps.Map = function (mapNode, options) {
         });
     };
 
-    // === Create marker with label ===
-    // https://github.com/printercu/google-maps-utility-library-v3-read-only/tree/master/markerwithlabel
-    self.addMarkerWithLabel = function (markerOpts) {
+    /**
+     * Create and add simple marker.
+     *
+     * @see https://developers.google.com/maps/documentation/javascript/markers
+     */
+    self.addMarker = function(position, title, icon) {
         var opts = {
-            map: self.map
+            position: position,
+            title: title,
+            icon: icon,
+            map: self.mapObject
+        };
+
+        var marker = new google.maps.Marker(opts);
+        self.markers.push(marker);
+
+        return marker;
+    };
+
+    /**
+     * Create and add marker with label.
+     *
+     * @see https://github.com/printercu/google-maps-utility-library-v3-read-only/tree/master/markerwithlabel
+     */
+    self.addMarkerWithLabel = function (position, markerOpts) {
+        var opts = {
+            position: position,
+            map: self.mapObject
         };
 
         $.extend(opts, markerOpts);